+++ subprojects/ide/src/integTest/groovy/org/gradle/plugins/ide/idea/CompositeBuildIdeaProjectIntegrationTest.groovy
@@ -21,7 +21,7 @@ import org.gradle.integtests.fixtures.ToBeFixedForConfigurationCache
 import org.gradle.integtests.fixtures.build.BuildTestFile
 import org.gradle.plugins.ide.fixtures.IdeaFixtures
 import org.gradle.test.fixtures.file.TestFile
-import org.gradle.util.ToBeImplemented
+import org.gradle.util.internal.ToBeImplemented
 import spock.lang.Issue
 
 /**
