+++ subprojects/language-java/src/integTest/groovy/org/gradle/java/compile/incremental/BaseJavaSourceIncrementalCompilationIntegrationTest.groovy
@@ -18,7 +18,7 @@ package org.gradle.java.compile.incremental
 
 import groovy.transform.NotYetImplemented
 import org.gradle.integtests.fixtures.CompiledLanguage
-import org.gradle.util.ToBeImplemented
+import org.gradle.util.internal.ToBeImplemented
 import spock.lang.Issue
 import spock.lang.Unroll
